Renamed Factory to just connector
