+++ src/SlackConnector.Tests.Unit/SlackConnector.Tests.Unit.csproj
@@ -87,12 +87,12 @@
     <Compile Include="Connections\Messaging\ChatMessengerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resources\ResourceManager.cs" />
-    <Compile Include="SlackConnectorFactoryTests\ConnectedStatusTests.cs" />
-    <Compile Include="SlackConnectorFactoryTests\HubsTests.cs" />
-    <Compile Include="SlackConnectorFactoryTests\MultipleConnections.cs" />
-    <Compile Include="SlackConnectorFactoryTests\TeamAndSelfTests.cs" />
-    <Compile Include="SlackConnectorFactoryTests\UsersTests.cs" />
-    <Compile Include="SlackConnectorFactoryTests\WebSocketTests.cs" />
+    <Compile Include="SlackConnectorTests\ConnectedStatusTests.cs" />
+    <Compile Include="SlackConnectorTests\HubsTests.cs" />
+    <Compile Include="SlackConnectorTests\MultipleConnections.cs" />
+    <Compile Include="SlackConnectorTests\TeamAndSelfTests.cs" />
+    <Compile Include="SlackConnectorTests\UsersTests.cs" />
+    <Compile Include="SlackConnectorTests\WebSocketTests.cs" />
     <Compile Include="SlackConnectionTests\Connect\ConnectedStatusTests.cs" />
     <Compile Include="SlackConnectionTests\InboundMessageTests.cs" />
     <Compile Include="SlackConnectionTests\Connect\MultipleConnections.cs" />
