+++ src/SlackConnector/SlackConnector.csproj
@@ -76,7 +76,7 @@
     <Compile Include="Exceptions\CommunicationException.cs" />
     <Compile Include="Exceptions\HandshakeException.cs" />
     <Compile Include="Exceptions\MissingChannelException.cs" />
-    <Compile Include="ISlackConnectorFactory.cs" />
+    <Compile Include="ISlackConnector.cs" />
     <Compile Include="Models\BotMessage.cs" />
     <Compile Include="EventHandlers\MessageReceivedEventHandler.cs" />
     <Compile Include="ISlackConnection.cs" />
@@ -104,7 +104,7 @@
     <Compile Include="Connections\Sockets\Messages\MessageSubType.cs" />
     <Compile Include="Connections\Sockets\Messages\MessageType.cs" />
     <Compile Include="Connections\Sockets\WebSocketClient.cs" />
-    <Compile Include="SlackConnectorFactory.cs" />
+    <Compile Include="SlackConnector.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
